feat: add state to table story for dynamic
